+++ Code/glimac/src/project/StaticObject.cpp
@@ -1,5 +1,5 @@
 #include "../include/project/StaticObject.hpp"
 
 
-StaticObject::StaticObject(int posX, int posY, int width, int height) : Object(posX, posY, width, height) { }
+StaticObject::StaticObject(int posX, int posY, float width, float height) : Object(posX, posY, width, height) { }
 StaticObject::StaticObject() { }
